MixPanel: handle SecurityErrorEvent and propogate it to callback
In case of a sandbox violation currently an unhandled securityError
occurs and since it cannot be handled externally to MixPanel it effects
the user experience badly. This can happen if api.mixpanel.com domain is
blocked by a company's firewall for instance.

Handle the SecurityErrorEvent on URLLoader and callback to provided
callback with the specific error as a second argument. Since this is
a new event handled and the first parameter remains 0 it should not affect
existing users, unless they checked arguments number.

The provided error can be used for better diagnostics.

Signed-off-by: Alon Levy <alonl@genomecompiler.com>